logout-service/requirements.txt

flask
pyjwt
python-dotenv
